pyproject.toml
[project]
name = "tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.0",
    "docx2txt>=0.9",
    "faiss-cpu>=1.13.2",
    "ipykernel>=7.1.0",
    "jq>=1.11.0",
    "langchain>=1.2.2",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "pandas>=2.3.3",
    "pymupdf>=1.27.1",
    "pypdf>=6.5.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "tiktoken>=0.12.0",
    "unstructured[docx]>=0.21.5",
]
